SUKU minor refactoring + element select fix
